3 authentification (#34)
* feat(auth): Implement authentication routes and components for sign-in, sign-up, and password reset

* feat(auth): Add user authentication and registration functionality with JWT support

* feat(auth): Refactor authentication components and routes to use 'login' and 'register' terminology; implement auth interceptor

* feat(recipe): Implement user-based recipe deletion and enhance recipe retrieval with user details

+++ api/src/main/java/com/a_table/service/RecipeService.java
@@ -1,9 +1,14 @@
 package com.a_table.service;
 
+import com.a_table.dto.Category;
 import com.a_table.dto.Recipe;
+import com.a_table.dto.User;
+import com.a_table.exception.InvalidCategoryException;
+import com.a_table.exception.RecipeCantBeDeleted;
 import com.a_table.exception.RecipeNotFoundException;
 import com.a_table.model.RecipeEntity;
 import com.a_table.model.RecipeStepEntity;
+import com.a_table.model.UserEntity;
 import com.a_table.repository.RecipeRepository;
 import com.a_table.utils.MappingService;
 import jakarta.annotation.Resource;
@@ -24,27 +29,55 @@ public class RecipeService {
     @Resource
     MappingService mappingService;
 
+    @Resource
+    UserService userService;
+
 
     public List<Recipe> getRecipes() {
         List<RecipeEntity> entities = recipeRepository.findAll();
-        entities.forEach(r -> {
-            if (r.getImage() != null && r.getImage().length > 0) {
-                r.setImageBase64("data:image/jpeg;base64," + Base64.getEncoder().encodeToString(r.getImage()));
+        entities.forEach(recipe -> {
+            if (recipe.getImage() != null && recipe.getImage().length > 0) {
+                recipe.setImageBase64("data:image/jpeg;base64," + Base64.getEncoder().encodeToString(recipe.getImage()));
             }
+
+            UserEntity user = new UserEntity();
+            user.setId(recipe.getUser().getId());
+            recipe.setUser(user);
         });
         return mappingService.convertListTo(entities, Recipe.class);
     }
 
     public Recipe getRecipe(Long id) {
-        RecipeEntity recipe = recipeRepository.findById(id).orElseThrow(RecipeNotFoundException::new);
-        if (recipe.getImage() != null && recipe.getImage().length > 0) {
-            recipe.setImageBase64("data:image/jpeg;base64," + Base64.getEncoder().encodeToString(recipe.getImage()));
+        RecipeEntity recipeEntity = recipeRepository.findById(id).orElseThrow(RecipeNotFoundException::new);
+        if (recipeEntity.getImage() != null && recipeEntity.getImage().length > 0) {
+            recipeEntity.setImageBase64("data:image/jpeg;base64," + Base64.getEncoder().encodeToString(recipeEntity.getImage()));
         }
-        recipe.getSteps().sort(Comparator.comparingInt(RecipeStepEntity::getStepNumber));
-        return mappingService.map(recipe, Recipe.class);
+        recipeEntity.getSteps().sort(Comparator.comparingInt(RecipeStepEntity::getStepNumber));
+        Recipe recipe = mappingService.map(recipeEntity, Recipe.class);
+        return recipe.toBuilder()
+                .user(User.builder()
+                        .id(recipe.getUser().getId())
+                        .firstName(recipe.getUser().getFirstName())
+                        .lastName(recipe.getUser().getLastName())
+                        .build())
+                .build();
     }
 
     public Recipe createRecipe(Recipe recipe) {
+        boolean categoryValid = Category.categoryValid(recipe.getCategory());
+        if (!categoryValid) {
+            throw new InvalidCategoryException();
+        }
+
+        User user = userService.getCurrentUser();
+
+        if (recipe.getImage() != null && recipe.getImage().startsWith("data:image")) {
+            String base64Image = recipe.getImage().split(",")[1];
+            byte[] imageBytes = Base64.getDecoder().decode(base64Image);
+            recipe.setImageBytes(imageBytes);
+        }
+
+        recipe.setUser(user);
         recipe.getIngredients().forEach(i -> i.setRecipe(recipe));
         recipe.getSteps().forEach(i -> i.setRecipe(recipe));
 
@@ -61,6 +94,11 @@ public class RecipeService {
 
     public void deleteRecipe(Long recipeId) {
         RecipeEntity recipeEntity = recipeRepository.findById(recipeId).orElseThrow(RecipeNotFoundException::new);
+        User user = userService.getCurrentUser();
+        if (!user.getId().equals(recipeEntity.getUser().getId())) {
+            throw new RecipeCantBeDeleted();
+        }
+
         recipeRepository.delete(recipeEntity);
     }
